part 2d complete

+++ part2/phonebook/src/App.jsx
@@ -3,6 +3,7 @@ import SearchFilter from './components/SearchFilter'
 import PersonForm from './components/PersonForm'
 import Persons from './components/Persons'
 import personService from './services/persons'
+import './index.css'
 
 const App = () => {
 
@@ -10,59 +11,80 @@ const App = () => {
   const [newName, setNewName] = useState('')
   const [newNumber, setNewNumber] = useState('')
   const [filterBy, setFilterBy] = useState('')
+  const [deleted, setDeleted] = useState(false)
+  const [updated, setUpdated] = useState(false)
 
   useEffect(() => {
       personService
       .getAllPersons()
-      .then(response => setPersons(response.data))
-  }, [])
+      .then(response => {
+        setPersons(response.data)
+        setDeleted(false)
+        setUpdated(false)
+        })
+  }, [deleted, updated])
 
   const handleOnChangeName = (event) => {
-    console.log('input', event.target.value)
+    //console.log('input', event.target.value)
     setNewName(event.target.value)
   }
 
   const handleOnChangeNumber = (event) => {
-    console.log('input', event.target.value)
+    //console.log('input', event.target.value)
     setNewNumber(event.target.value)
   }
 
   const addName = (event) => {
     event.preventDefault()
     
-    const nameExists = persons.find(x => {
-      console.log('x.name',x.name)
-      console.log('newName', newName)
-      return x.name === newName
-  })
+    const nameExists = persons.find(x => x.name === newName)
 
     console.log('nameExists', nameExists)
     if (nameExists !== undefined) {
-      alert(`${newName} is already added to phonebook`)
+      if (nameExists.number === newNumber) {
+        alert(`${newName} is already added to phonebook`)
+        return
+      } else {
+        if (window.confirm(`${newName} is already added to phonebook. Replace new number with old one ?`)) {
+          
+          const newPerson = {
+            name: newName,
+            number: newNumber,
+            id: nameExists.id
+          }
+
+          personService
+          .updatePerson(newPerson)
+          .then(response => {
+            setUpdated(true)
+            console.log('updated response', response.data)
+          })
+        }
+      }
+      
+
     } else { 
       const newPerson = {
         name: newName,
         number: newNumber
       }
 
-      useEffect, (
         personService
         .createPerson(newPerson)
         .then(response => {
            setPersons(persons.concat(response.data))
            setNewName('')
            setNewNumber('')
-        }),[])
-
+        })
     }
   }
 
   const handleDelete = selectedPerson => {
-    console.log('selectedPerson', selectedPerson)
-    personService.deletePerson(selectedPerson)
-    .then(response)
+    if (window.confirm(`Delete ${selectedPerson.name} ?`)){
+      personService.deletePerson(selectedPerson)
+      .then(response => setDeleted(true))
+    } 
   }
-
 
 
   const filterNames = (event) => {
@@ -82,7 +104,7 @@ const App = () => {
 
   return (
     <div>
-      <h2>Phonebook</h2>
+      <h1>Phonebook</h1>
       <SearchFilter filterBy={filterBy} filterNames={filterNames} />
       <h2>add new</h2>
       <PersonForm
